requirements.txt
fastapi
uvicorn[standard]
asyncpg
sqlalchemy[asyncio]
pydantic
python-dotenv
